requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
opencv-python
tensorflow
pyyaml